experiment with splitting engine.go into smaller files
