+++ client/engine/handle_chain_event.go
@@ -1,0 +1,30 @@
+package engine
+
+import "github.com/statechannels/go-nitro/client/engine/chainservice"
+
+// handleChainEvent handles a Chain Event from the blockchain.
+// It:
+//   - reads an objective from the store,
+//   - generates an updated objective, and
+//   - attempts progress.
+func (e *Engine) handleChainEvent(chainEvent chainservice.Event) (EngineEvent, error) {
+	defer e.metrics.RecordFunctionDuration()()
+	e.logger.Printf("handling chain event %v", chainEvent)
+	objective, ok := e.store.GetObjectiveByChannelId(chainEvent.ChannelID())
+	if !ok {
+		// TODO: Right now the chain service returns chain events for ALL channels even those we aren't involved in
+		// for now we can ignore channels we aren't involved in
+		// in the future the chain service should allow us to register for specific channels
+		return EngineEvent{}, nil
+	}
+
+	eventHandler, ok := objective.(chainservice.ChainEventHandler)
+	if !ok {
+		return EngineEvent{}, &ErrUnhandledChainEvent{event: chainEvent, objective: objective, reason: "objective does not handle chain events"}
+	}
+	updatedEventHandler, err := eventHandler.UpdateWithChainEvent(chainEvent)
+	if err != nil {
+		return EngineEvent{}, err
+	}
+	return e.attemptProgress(updatedEventHandler)
+}
