+++ client/engine/handle_message.go
@@ -1,0 +1,199 @@
+package engine
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/statechannels/go-nitro/channel/consensus_channel"
+	"github.com/statechannels/go-nitro/channel/state"
+	"github.com/statechannels/go-nitro/protocols"
+	"github.com/statechannels/go-nitro/protocols/directdefund"
+	"github.com/statechannels/go-nitro/protocols/directfund"
+	"github.com/statechannels/go-nitro/protocols/virtualdefund"
+	"github.com/statechannels/go-nitro/protocols/virtualfund"
+)
+
+// handleMessage handles a Message from a peer go-nitro Wallet.
+// It:
+//   - reads an objective from the store,
+//   - generates an updated objective,
+//   - attempts progress on the target Objective,
+//   - attempts progress on related objectives which may have become unblocked.
+func (e *Engine) handleMessage(message protocols.Message) (EngineEvent, error) {
+	defer e.metrics.RecordFunctionDuration()()
+
+	e.logger.Printf("Handling inbound message %+v", protocols.SummarizeMessage(message))
+	allCompleted := EngineEvent{}
+
+	for _, entry := range message.SignedStates() {
+
+		objective, err := e.getOrCreateObjective(entry.ObjectiveId, entry.Payload)
+		if err != nil {
+			return EngineEvent{}, err
+		}
+
+		if objective.GetStatus() == protocols.Unapproved {
+			e.logger.Printf("Policymaker is %+v", e.policymaker)
+			if e.policymaker.ShouldApprove(objective) {
+				objective = objective.Approve()
+
+				ddfo, ok := objective.(*directdefund.Objective)
+				if ok {
+					// If we just approved a direct defund objective, destroy the consensus channel to prevent it being used (a Channel will now take over governance)
+					e.store.DestroyConsensusChannel(ddfo.C.Id)
+				}
+			} else {
+				objective, sideEffects := objective.Reject()
+				err = e.store.SetObjective(objective)
+				if err != nil {
+					return EngineEvent{}, err
+				}
+
+				allCompleted.CompletedObjectives = append(allCompleted.CompletedObjectives, objective)
+				err = e.executeSideEffects(sideEffects)
+				// An error would mean we failed to send a message. But the objective is still "completed".
+				// So, we should return allCompleted even if there was an error.
+				return allCompleted, err
+			}
+		}
+
+		if objective.GetStatus() == protocols.Completed {
+			e.logger.Printf("Ignoring payload for complected objective  %s", objective.Id())
+			continue
+		}
+		if objective.GetStatus() == protocols.Rejected {
+			e.logger.Printf("Ignoring payload for rejected objective  %s", objective.Id())
+			continue
+		}
+
+		event := protocols.ObjectiveEvent{
+			ObjectiveId:    entry.ObjectiveId,
+			SignedProposal: consensus_channel.SignedProposal{},
+			SignedState:    entry.Payload,
+		}
+		updatedObjective, err := objective.Update(event)
+		if err != nil {
+			return EngineEvent{}, err
+		}
+
+		progressEvent, err := e.attemptProgress(updatedObjective)
+		if err != nil {
+			return EngineEvent{}, err
+		}
+		allCompleted.CompletedObjectives = append(allCompleted.CompletedObjectives, progressEvent.CompletedObjectives...)
+
+		if err != nil {
+			return EngineEvent{}, err
+		}
+
+	}
+
+	for _, entry := range message.SignedProposals() {
+		e.logger.Printf("handling proposal %+v", protocols.SummarizeProposal(entry.ObjectiveId, entry.Payload))
+		objective, err := e.store.GetObjectiveById(entry.ObjectiveId)
+		if err != nil {
+			return EngineEvent{}, err
+		}
+		if objective.GetStatus() == protocols.Completed {
+			e.logger.Printf("Ignoring payload for complected objective  %s", objective.Id())
+			continue
+		}
+
+		event := protocols.ObjectiveEvent{
+			ObjectiveId:    entry.ObjectiveId,
+			SignedProposal: entry.Payload,
+			SignedState:    state.SignedState{},
+		}
+		updatedObjective, err := objective.Update(event)
+		if err != nil {
+			return EngineEvent{}, err
+		}
+
+		progressEvent, err := e.attemptProgress(updatedObjective)
+		if err != nil {
+			return EngineEvent{}, err
+		}
+
+		allCompleted.CompletedObjectives = append(allCompleted.CompletedObjectives, progressEvent.CompletedObjectives...)
+
+		if err != nil {
+			return EngineEvent{}, err
+		}
+
+	}
+
+	for _, entry := range message.RejectedObjectives() {
+		objective, err := e.store.GetObjectiveById(entry.ObjectiveId)
+
+		if err != nil {
+			return EngineEvent{}, err
+		}
+		if objective.GetStatus() == protocols.Rejected {
+			e.logger.Printf("Ignoring payload for rejected objective  %s", objective.Id())
+			continue
+		}
+
+		// we are rejecting due to a counterparty message notifying us of their rejection. We
+		// do not need to send a message back to that counterparty, and furthermore we assume that
+		// counterparty has already notified all other interested parties. We can therefore ignore the side effects
+		objective, _ = objective.Reject()
+		err = e.store.SetObjective(objective)
+		if err != nil {
+			return EngineEvent{}, err
+		}
+
+		allCompleted.CompletedObjectives = append(allCompleted.CompletedObjectives, objective)
+	}
+
+	for _, voucher := range message.Vouchers() {
+
+		// TODO: return the amount we paid?
+		_, err := e.vm.Receive(voucher)
+
+		allCompleted.ReceivedVouchers = append(allCompleted.ReceivedVouchers, voucher)
+		if err != nil {
+			return EngineEvent{}, fmt.Errorf("error accepting payment voucher: %w", err)
+		}
+
+	}
+	return allCompleted, nil
+
+}
+
+// constructObjectiveFromMessage Constructs a new objective (of the appropriate concrete type) from the supplied message.
+func (e *Engine) constructObjectiveFromMessage(id protocols.ObjectiveId, ss state.SignedState) (protocols.Objective, error) {
+	defer e.metrics.RecordFunctionDuration()()
+
+	switch {
+	case directfund.IsDirectFundObjective(id):
+		dfo, err := directfund.ConstructFromState(false, ss.State(), *e.store.GetAddress())
+
+		return &dfo, err
+	case virtualfund.IsVirtualFundObjective(id):
+		vfo, err := virtualfund.ConstructObjectiveFromState(ss.State(), false, *e.store.GetAddress(), e.store.GetConsensusChannel)
+		if err != nil {
+			return &virtualfund.Objective{}, fmt.Errorf("could not create virtual fund objective from message: %w", err)
+		}
+		err = e.registerPaymentChannel(vfo)
+		if err != nil {
+			return &virtualfund.Objective{}, fmt.Errorf("could not register channel with payment/receipt manager: %w", err)
+		}
+		return &vfo, nil
+	case virtualdefund.IsVirtualDefundObjective(id):
+		vdfo, err := virtualdefund.ConstructObjectiveFromState(ss.State(), false, *e.store.GetAddress(), e.store.GetChannelById, e.store.GetConsensusChannel)
+		if err != nil {
+			return &virtualfund.Objective{}, fmt.Errorf("could not create virtual fund objective from message: %w", err)
+		}
+		return &vdfo, nil
+	case directdefund.IsDirectDefundObjective(id):
+		ddfo, err := directdefund.ConstructObjectiveFromState(ss.State(), false, e.store.GetConsensusChannelById)
+		if err != nil {
+			return &directdefund.Objective{}, fmt.Errorf("could not create direct defund objective from message: %w", err)
+		}
+		return &ddfo, nil
+
+	default:
+		return &directfund.Objective{}, errors.New("cannot handle unimplemented objective type")
+	}
+
+}
