+++ client/engine/handle_payment_request.go
@@ -1,0 +1,34 @@
+package engine
+
+import (
+	"fmt"
+
+	"github.com/statechannels/go-nitro/payments"
+	"github.com/statechannels/go-nitro/protocols"
+)
+
+// handleObjectiveRequest handles an PaymentRequest (triggered by a client API call).
+// It prepares and dispatches a payment message to the counterparty.
+func (e *Engine) handlePaymentRequest(request PaymentRequest) error {
+	if (request == PaymentRequest{}) {
+		panic("tried to handle nil payment request")
+	}
+	cId := request.ChannelId
+	voucher, err := e.vm.Pay(
+		cId,
+		request.Amount,
+		*e.store.GetChannelSecretKey())
+	if err != nil {
+		return fmt.Errorf("handleAPIEvent: Error making payment: %w", err)
+	}
+	c, ok := e.store.GetChannelById(cId)
+	if !ok {
+		return fmt.Errorf("handleAPIEvent: Could not get channel from the store %s", cId)
+	}
+	payer, payee := payments.GetPayer(c.Participants), payments.GetPayee(c.Participants)
+	if payer != *e.store.GetAddress() {
+		return fmt.Errorf("handleAPIEvent: Not the sender in channel %s", cId)
+	}
+	se := protocols.SideEffects{MessagesToSend: protocols.CreateVoucherMessage(voucher, payee)}
+	return e.executeSideEffects(se)
+}
