+++ client/engine/handle_proposal.go
@@ -1,0 +1,101 @@
+package engine
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/statechannels/go-nitro/channel/consensus_channel"
+	"github.com/statechannels/go-nitro/channel/state"
+	"github.com/statechannels/go-nitro/client/engine/store"
+	"github.com/statechannels/go-nitro/protocols"
+	"github.com/statechannels/go-nitro/protocols/directfund"
+	"github.com/statechannels/go-nitro/protocols/virtualdefund"
+	"github.com/statechannels/go-nitro/protocols/virtualfund"
+)
+
+// handleProposal handles a Proposal returned to the engine from
+// a running ledger channel by pulling its corresponding objective
+// from the store and attempting progress.
+func (e *Engine) handleProposal(proposal consensus_channel.Proposal) (EngineEvent, error) {
+	defer e.metrics.RecordFunctionDuration()()
+
+	id := getProposalObjectiveId(proposal)
+	obj, err := e.store.GetObjectiveById(id)
+	if err != nil {
+		return EngineEvent{}, err
+	}
+	return e.attemptProgress(obj)
+}
+
+// getProposalObjectiveId returns the objectiveId for a proposal.
+func getProposalObjectiveId(p consensus_channel.Proposal) protocols.ObjectiveId {
+	switch p.Type() {
+	case consensus_channel.AddProposal:
+		{
+			const prefix = virtualfund.ObjectivePrefix
+			channelId := p.ToAdd.Guarantee.Target().String()
+			return protocols.ObjectiveId(prefix + channelId)
+
+		}
+	case consensus_channel.RemoveProposal:
+		{
+			const prefix = virtualdefund.ObjectivePrefix
+			channelId := p.ToRemove.Target.String()
+			return protocols.ObjectiveId(prefix + channelId)
+
+		}
+	default:
+		{
+			panic("invalid proposal type")
+		}
+	}
+}
+
+// spawnConsensusChannelIfDirectFundObjective will attempt to create and store a ConsensusChannel derived from the supplied Objective if it is a directfund.Objective.
+//
+// The associated Channel will remain in the store.
+func (e Engine) spawnConsensusChannelIfDirectFundObjective(crankedObjective protocols.Objective) error {
+	defer e.metrics.RecordFunctionDuration()()
+
+	if dfo, isDfo := crankedObjective.(*directfund.Objective); isDfo {
+		c, err := dfo.CreateConsensusChannel()
+		if err != nil {
+			return fmt.Errorf("could not create consensus channel for objective %s: %w", crankedObjective.Id(), err)
+		}
+		err = e.store.SetConsensusChannel(c)
+		if err != nil {
+			return fmt.Errorf("could not store consensus channel for objective %s: %w", crankedObjective.Id(), err)
+		}
+		// Destroy the channel since the consensus channel takes over governance:
+		e.store.DestroyChannel(c.Id)
+	}
+	return nil
+}
+
+// getOrCreateObjective retrieves the objective from the store. if the objective does not exist, it creates the objective using the supplied signed state, and stores it in the store
+func (e *Engine) getOrCreateObjective(id protocols.ObjectiveId, ss state.SignedState) (protocols.Objective, error) {
+	defer e.metrics.RecordFunctionDuration()()
+
+	objective, err := e.store.GetObjectiveById(id)
+
+	if err == nil {
+		return objective, nil
+	} else if errors.Is(err, store.ErrNoSuchObjective) {
+
+		newObj, err := e.constructObjectiveFromMessage(id, ss)
+
+		if err != nil {
+			return nil, fmt.Errorf("error constructing objective from message: %w", err)
+		}
+		e.metrics.RecordObjectiveStarted(newObj.Id())
+		err = e.store.SetObjective(newObj)
+		if err != nil {
+			return nil, fmt.Errorf("error setting objective in store: %w", err)
+		}
+		e.logger.Printf("Created new objective from  message %s", newObj.Id())
+		return newObj, nil
+
+	} else {
+		return nil, fmt.Errorf("unexpected error getting/creating objective %s: %w", id, err)
+	}
+}
