+++ client/engine/types.go
@@ -1,0 +1,45 @@
+package engine
+
+import (
+	"fmt"
+	"math/big"
+
+	"github.com/statechannels/go-nitro/client/engine/chainservice"
+	"github.com/statechannels/go-nitro/payments"
+	"github.com/statechannels/go-nitro/protocols"
+	"github.com/statechannels/go-nitro/types"
+)
+
+// PaymentRequest represents a request from the API to make a payment using a channel
+type PaymentRequest struct {
+	ChannelId types.Destination
+	Amount    *big.Int
+}
+
+// EngineEvent is a struct that contains a list of changes caused by handling a message/chain event/api event
+type EngineEvent struct {
+	// These are objectives that are now completed
+	CompletedObjectives []protocols.Objective
+	// These are objectives that have failed
+	FailedObjectives []protocols.ObjectiveId
+	// ReceivedVouchers are vouchers we've received from other participants
+	ReceivedVouchers []payments.Voucher
+}
+
+type CompletedObjectiveEvent struct {
+	Id protocols.ObjectiveId
+}
+
+// Response is the return type that asynchronous API calls "resolve to". Such a call returns a go channel of type Response.
+type Response struct{}
+
+// ErrUnhandledChainEvent is an engine error when the the engine cannot process a chain event
+type ErrUnhandledChainEvent struct {
+	event     chainservice.Event
+	objective protocols.Objective
+	reason    string
+}
+
+func (uce *ErrUnhandledChainEvent) Error() string {
+	return fmt.Sprintf("chain event %#v could not be handled by objective %#v due to: %s", uce.event, uce.objective, uce.reason)
+}
